restructuring project code [ch #168282619]
