+++ api/controllers/session.js
@@ -1,12 +1,9 @@
 import Sessions from '../models/session';
-import Auth from '../middleware/auth';
 
 const SessionsData = [];
 
 class SessionController{
     static createSession(req, res){
-        // const user = Auth.verifyUser(req, res);
-        // return res.status(201).send(user)
         const sessionId = SessionsData.length + 1;
         const status = "pending";
         const menteeId = req.params.token;
