+++ api/controllers/user.js
@@ -3,79 +3,43 @@ import Sessions from '../controllers/session';
 import dotenv from 'dotenv';
 import validate from '../helpers/helper';
 
-const users = [];
+
+
+
 dotenv.config();
+
+const userClass = User.User
+const userList = User.users
 
 class UserController{
     static RegisterUser(req, res){
-        const id = users.length +1;
-        const passInput = req.body.password
-        if (passInput.length < 8){
-            return res.status(400).send({
-                status: 400,
-                error: "Password too short"
-            });
-        }
-        var pass1 = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{6,20}$/;
-        const check2 = passInput.match(pass1);
-        if(!check2){
-            
-        }
-        const password = validate.hashPassword(passInput)
-        // const isMentor = true;
-        const userEmail = users.find(user => user.email === req.body.email);
-        if(!validate.isValidEmail(req.body.email)){
-            return res.status(400).json({
-                'status': 400,
-                'error': 'The email you provided is not valid'
-            })
-        }
-        if (userEmail){
-            return res.status(409).send({
-                status: 409,
-                error: "User already registered with this email!"
-            }) 
-        }
-        const user = new User(
+        const id = userList.length +1;
+        const user = new userClass(
             id, req.body.firstName, req.body.lastName,
-            req.body.email, password, req.body.address, req.body.occupation,
+            req.body.email, req.AuthorizePassword, req.body.address, req.body.occupation,
             req.body.expertise
         )
         const token = validate.generateToken(user.email)
-        users.push(user);
+        userList.push(user);
         return res.status(201).json({
             status: 201,
             message: 'User created successfully',
             data:{
                 token: token,
                 id: user.id
-                // status: user.status
             }
         })
     }
 
 
     static UserLogin(req, res){
-        const loginUser = users.find(user => user.email === req.body.email);
-        if(!loginUser){
-            return res.status(404).send({
-                status: 404,
-                error: `User with  was not found`
-            });
-        }
-        const passwordCompared = validate.comparePassword(loginUser.password, req.body.password);
-        if(!passwordCompared){
-            return res.status(401).send({
-                status: 401,
-                error: "Login was denied"
-            });
-        }
-        const token = validate.generateToken(loginUser.id, loginUser.email)
+        
+        const token = validate.generateToken(req.userId, req.email);
         return res.status(200).send({
             status: 200,
             data: {
                 token: token,
-                id: loginUser.id,
+                id: req.userId,
                 message: "You are logged in successfully"
             } 
         });
@@ -99,7 +63,7 @@ class UserController{
 
     static GetOneMentor(req, res){
         const oneMentor = req.body.id;
-        const availableMentor = users.find(user => user.isMentor == "true", user => user.id === oneMentor);
+        const availableMentor = userList.find(user => user.isMentor == "true", user => user.id === oneMentor);
         if(!availableMentor){    
             return res.status(404).send({
                 status: 404,
@@ -182,4 +146,4 @@ class UserController{
 
 
 
-export default {UserController, users};
+export default {UserController};
