+++ api/tests/1_users.js
@@ -1,22 +1,21 @@
 import chai, {expect} from 'chai';
 import chaiHttp from 'chai-http';
+import userDetails from './mocks/mocks';
 import app from '../server';
 
 
 chai.should();
 chai.use(chaiHttp);
 
-let adminToken;
+let userToken;
+const user = userDetails.users
 
-describe('Testing Admin Routes', () =>{
-    it('should signup a new admin', (done) => {
+describe('Testing routes for the user', () =>{
+    it('should signup a new user', (done) => {
         chai
             .request(app)
-            .post('/api/v1/auth/admin/signup')
-            .send({
-                "email":"test@admin.com",
-                "password":"test123"
-            })
+            .post('/api/v1/auth/signup')
+            .send(user[0])
             .then((res) => {
                 expect(res.body).to.be.an('object');
                 expect(res.body).to.have.property('status');
@@ -26,14 +25,44 @@ describe('Testing Admin Routes', () =>{
             })
             .catch(err => done(err));
     });
-
-    it('admin already created', (done) => {
+    it('A user can not create two accounts with same email', (done) => {
         chai
             .request(app)
-            .post('/api/v1/auth/admin/signup')
+            .post('/api/v1/auth/signup')
+            .send(user[0])
+            .then((res) => {
+                expect(res.body).to.be.an('object');
+                expect(res.body).to.have.property('status');
+                expect(res.body).to.have.property('error');
+                done();
+            })
+            .catch(err => done(err));
+    });
+    it('should login a user', (done) => {
+        chai
+            .request(app)
+            .post('/api/v1/auth/login')
             .send({
-                "email":"test@admin.com",
-                "password":"test123"
+                "email": user[0]['email'],
+                "password": user[0]['password']
+            })
+            .then((res) => {
+                userToken = res.body.data['token'];
+                expect(res.body).to.be.an('object');
+                expect(res.body).to.have.property('status');
+                expect(res.body).to.have.property('data');
+                expect(res.body).to.have.property('data').to.be.an('object');
+                done();
+            })
+            .catch(err => done(err));
+    });
+    it('login user not found', (done) => {
+        chai
+            .request(app)
+            .post('/api/v1/auth/login')
+            .send({
+                "email": "twothree@test.com",
+                "password": "test123"
             })
             .then((res) => {
                 expect(res.body).to.be.an('object');
@@ -43,33 +72,13 @@ describe('Testing Admin Routes', () =>{
             })
             .catch(err => done(err));
     });
-
-    it('should login admin', (done) => {
+    it('login is denied', (done) => {
         chai
             .request(app)
-            .post('/api/v1/auth/admin/login')
+            .post('/api/v1/auth/login')
             .send({
-                "email":"test@admin.com",
-                "password":"test123"
-            })
-            .then((res) => {
-                adminToken = res.body.data['token'];
-                expect(res.body).to.be.an('object');
-                expect(res.body).to.have.property('status');
-                expect(res.body).to.have.property('data');
-                expect(res.body).to.have.property('data').to.be.an('object');
-                done();
-            })
-            .catch(err => done(err));
-    })
-
-    it('admin user not found', (done) => {
-        chai
-            .request(app)
-            .post('/api/v1/auth/admin/login')
-            .send({
-                "email":"testtest@admin.com",
-                "password":"test123"
+                "email": "two@test.com",
+                "password": "test1231"
             })
             .then((res) => {
                 expect(res.body).to.be.an('object');
@@ -78,23 +87,6 @@ describe('Testing Admin Routes', () =>{
                 done();
             })
             .catch(err => done(err));
-    })
-
-    it('login denied', (done) => {
-        chai
-            .request(app)
-            .post('/api/v1/auth/admin/login')
-            .send({
-                "email":"test@admin.com",
-                "password":"test1231"
-            })
-            .then((res) => {
-                expect(res.body).to.be.an('object');
-                expect(res.body).to.have.property('status');
-                expect(res.body).to.have.property('error');
-                done();
-            })
-            .catch(err => done(err));
-    })
+    });
     
 })
