+++ api/tests/mocks/mocks.js
@@ -1,0 +1,20 @@
+const users = [
+    {
+        "firstName": "Davies",
+        "lastName": "Wabuluka",
+        "email": "two@test.com",
+        "password": "test@123",
+        "address": "nalumunye",
+        "bio": "a good man",
+        "occupation": "teacher",
+        "expertise": "cooking"
+    }
+];
+const admin = [
+    {
+        "email":"test@admin.com",
+        "password":"test123"
+    }
+];
+
+export default {users, admin};
